requirements.txt
Flask==3.1.2
pymongo==4.10.1
python-dotenv==1.0.1
PyJWT==2.10.0
requests==2.32.3
beautifulsoup4==4.14.3
flask-cors==5.0.0
gunicorn==23.0.0
bcrypt==4.2.1
flask-limiter==3.9.2
elasticsearch>=8.0.0,<9.0.0
